tooling: Cleaup dep tools (#19816)
